% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Accessors.R
\name{GetRNA}
\alias{GetRNA}
\alias{GetATAC}
\alias{GetCellMetadata}
\alias{GetGeneAnnotation}
\alias{GetVariableGenes}
\alias{GetParameters}
\alias{GetCellEmbeddings}
\alias{GetRNA,BayesCNet-method}
\alias{GetATAC,BayesCNet-method}
\alias{GetCellMetadata,BayesCNet-method}
\alias{GetGeneAnnotation,BayesCNet-method}
\alias{GetVariableGenes,BayesCNet-method}
\alias{GetParameters,BayesCNet-method}
\alias{GetCellEmbeddings,BayesCNet-method}
\title{Basic accessors for BayesCNet objects}
\usage{
GetRNA(x)

GetATAC(x)

GetCellMetadata(x)

GetGeneAnnotation(x)

GetVariableGenes(x)

GetParameters(x)

GetCellEmbeddings(x)

\S4method{GetRNA}{BayesCNet}(x)

\S4method{GetATAC}{BayesCNet}(x)

\S4method{GetCellMetadata}{BayesCNet}(x)

\S4method{GetGeneAnnotation}{BayesCNet}(x)

\S4method{GetVariableGenes}{BayesCNet}(x)

\S4method{GetParameters}{BayesCNet}(x)

\S4method{GetCellEmbeddings}{BayesCNet}(x)
}
\arguments{
\item{x}{A BayesCNet object}
}
\value{
The requested data
}
\description{
Methods to access data from BayesCNet objects
}
