% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{AddCellTypeHierarchy}
\alias{AddCellTypeHierarchy}
\title{Add cell type hierarchy to BayesCNet object}
\usage{
AddCellTypeHierarchy(object, edges, lambda = 1)
}
\arguments{
\item{object}{A BayesCNet object}

\item{edges}{Character vector of edges in format c("parent1", "child1", "parent2", "child2", ...)}

\item{lambda}{Decay parameter for tree-based distance (default: 1.0)}
}
\value{
Updated BayesCNet object with hierarchy
}
\description{
Add cell type hierarchy to BayesCNet object
}
\examples{
\dontrun{
edges <- c("HSPC", "T Cells", "HSPC", "B Cells", "T Cells", "CD4 T", "T Cells", "CD8 T")
bcnet <- AddCellTypeHierarchy(bcnet, edges, lambda = 1.0)
}
}
