% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{AddGeneAnnotation}
\alias{AddGeneAnnotation}
\title{Add gene annotation to BayesCNet object}
\usage{
AddGeneAnnotation(object, annotation, species = NULL)
}
\arguments{
\item{object}{A BayesCNet object}

\item{annotation}{Data frame with columns: gene, chr, tss}

\item{species}{Species for automatic annotation (not implemented yet)}
}
\value{
Updated BayesCNet object with gene annotation
}
\description{
Add gene annotation to BayesCNet object
}
\examples{
\dontrun{
# annotation should have columns: gene, chr, tss
bcnet <- AddGeneAnnotation(bcnet, gene_annotation_df)
}
}
