% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{aggregate_single_celltype}
\alias{aggregate_single_celltype}
\title{Aggregate a single cell type}
\usage{
aggregate_single_celltype(
  rna_data,
  atac_data,
  cell_coord,
  k_neigh = 50,
  max_overlap = 0.8,
  seed = 123,
  verbose = FALSE
)
}
\arguments{
\item{rna_data}{RNA count matrix for one cell type}

\item{atac_data}{ATAC count matrix for one cell type}

\item{cell_coord}{Cell coordinates for this cell type}

\item{k_neigh}{Number of neighbors}

\item{max_overlap}{Maximum overlap allowed}

\item{seed}{Random seed}

\item{verbose}{Print messages}
}
\value{
List with aggregated data
}
\description{
Aggregate a single cell type
}
\keyword{internal}
