% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BayesCNet-class}
\alias{BayesCNet-class}
\title{The BayesCNet class}
\description{
The BayesCNet class
}
\section{Slots}{

\describe{
\item{\code{RNA}}{RNA count matrix (genes x cells)}

\item{\code{ATAC}}{ATAC count matrix (peaks x cells)}

\item{\code{cell_metadata}}{DataFrame with cell metadata}

\item{\code{gene_annotation}}{DataFrame with gene TSS information}

\item{\code{cell_type_hierarchy}}{List containing hierarchy information}

\item{\code{aggregated_data}}{List containing aggregated cell data}

\item{\code{variable_genes}}{Character vector of selected genes}

\item{\code{cell_embeddings}}{Matrix of cell embeddings (cells x 2)}

\item{\code{embedding_type}}{Character string indicating embedding type}

\item{\code{inference_results}}{DataFrame with inference results}

\item{\code{parameters}}{List of parameters used}

\item{\code{project_name}}{Character string for project name}
}}

