% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBayesCNet.R
\name{CreateBayesCNet}
\alias{CreateBayesCNet}
\title{Create a BayesCNet object from Seurat}
\usage{
CreateBayesCNet(
  seurat_object,
  rna_assay = "RNA",
  atac_assay = "ATAC",
  cell_type_col = "celltype",
  project_name = NULL
)
}
\arguments{
\item{seurat_object}{A Seurat object with RNA and ATAC assays}

\item{rna_assay}{Name of RNA assay in Seurat object}

\item{atac_assay}{Name of ATAC assay in Seurat object}

\item{cell_type_col}{Column name in metadata for cell types}

\item{project_name}{Character string for project name}
}
\value{
A BayesCNet object
}
\description{
Create a BayesCNet object from Seurat
}
\examples{
\dontrun{
bcnet <- CreateBayesCNet(pbmc_seurat)
}
}
