% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{estimateSizeFactorsForMatrix}
\alias{estimateSizeFactorsForMatrix}
\title{Estimate size factors for matrix normalization}
\usage{
estimateSizeFactorsForMatrix(mat)
}
\arguments{
\item{mat}{Count matrix}
}
\value{
Numeric vector of size factors
}
\description{
Estimate size factors for matrix normalization
}
\keyword{internal}
