% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{find_overlapping_coordinates}
\alias{find_overlapping_coordinates}
\title{Find overlapping genomic coordinates}
\usage{
find_overlapping_coordinates(peak_list, target_region)
}
\arguments{
\item{peak_list}{Character vector of peaks in format "chr_start_end"}

\item{target_region}{Target region in same format}
}
\value{
Character vector of overlapping peaks
}
\description{
Find overlapping genomic coordinates
}
\keyword{internal}
