% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessors.R
\name{IsReadyForInference}
\alias{IsReadyForInference}
\title{Check if BayesCNet object is ready for inference}
\usage{
IsReadyForInference(object)
}
\arguments{
\item{object}{A BayesCNet object}
}
\value{
Logical indicating if object is ready
}
\description{
Check if BayesCNet object is ready for inference
}
