% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,BayesCNet-method}
\alias{show,BayesCNet-method}
\title{Show method for BayesCNet}
\usage{
\S4method{show}{BayesCNet}(object)
}
\arguments{
\item{object}{A BayesCNet object}
}
\description{
Show method for BayesCNet
}
