% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessors.R
\name{summary,BayesCNet-method}
\alias{summary,BayesCNet-method}
\title{Summary method for BayesCNet}
\usage{
\S4method{summary}{BayesCNet}(object)
}
\arguments{
\item{object}{A BayesCNet object}
}
\description{
Summary method for BayesCNet
}
